requirements.txt
# django
Django==1.8.1

# django apps
django-floppyforms==1.4.0
#South==0.7.6

# general libs
pillow==2.8.1
bleach==1.4.1
pytz

#search
Whoosh==2.7.0
django-haystack==2.3.1

# queue
redis
hiredis
rq
django-rq

# dev stuff
django-extensions==1.5.3
django-debug-toolbar==1.3.0
django-nose==1.4
django-taggit==0.14.0
ipython
--allow-external python-graph-core
--allow-unverified python-graph-core
python-graph-core==1.8.2
# fabric
# fabric-virtualenv
